refactor: modularize timeline builder
