+++ src/timeline/constants.ts
@@ -1,0 +1,4 @@
+export const LINE_WIPE_DURATION = 0.5;
+export const DEFAULT_CHARS_PER_LINE = 40;
+export const MIN_FONT_SIZE = 24;
+export const APPROX_CHAR_WIDTH_RATIO = 0.56;
